Atemschutzüberwachung/requirements.txt
Flask>=2.0
requests
python-dotenv
